requirements.txt
fastapi
uvicorn
pandas
python-multipart
scikit-learn
pyarrow
numpy
scikit-learn
boto3
openai
python-dotenv